% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.KFPLS.R
\name{predict.KFPLS}
\alias{predict.KFPLS}
\title{Prediction by KFPLS}
\usage{
\method{predict}{KFPLS}(object, newdata, ...)
}
\arguments{
\item{object}{A KFPLS object obtained from \code{\link[KFPLS]{KFPLS}}.}

\item{newdata}{An \code{array} with three indices denoting the new observations of the functional predictors. The (i, j, k)-th element of it corresponds to the measurment of the i-th subject for the k-th functional predictor at j-th observation grid.}

\item{...}{Not used.}
}
\value{
A \code{vector} denoting the prediction of the scalar response.
}
\description{
Prediction of the scalar response by KFPLS.
}
